# Revert pointwise concat heuristics