--- a/torch/_inductor/config.py
+++ b/torch/_inductor/config.py
@@ -277,6 +277,7 @@ max_fusion_size = 64
 
 # max number of inputs to generate cat as a pointwise op with masked laods
 max_pointwise_cat_inputs = 8
+max_pointwise_cat_inputs = 4
 
 # replace small reductions with pointwise, disable with `= 1`
 unroll_reductions_threshold = 8