@eellison 2024-02-14
@xuzhao9 the 10% regression was with 128 - I have already lowered it down to 8. Can you check if the regression still occurs?